+++ lib/polyfill/v2_4/string_io.rb
@@ -1,0 +1,9 @@
+require 'polyfill/v2_4/string_io/instance'
+
+module Polyfill
+  module V2_4
+    module StringIO
+      include Instance
+    end
+  end
+end